unify encrypted vote-related stuff
